+++ commands/TranslationController.php
@@ -3,7 +3,6 @@
 namespace humhub\modules\translation\commands;
 
 use humhub\modules\translation\models\Languages;
-use Yii;
 use yii\base\Exception;
 use yii\console\Controller;
 
@@ -22,6 +21,15 @@ class TranslationController extends Controller
         }
     }
 
+    /**
+     * Rename category, use * for old category param if need to rename all old categories to new one, or separate old categories with comma
+     *
+     * @param $moduleId
+     * @param $oldCategory
+     * @param $newCategory
+     * @return void
+     * @throws Exception
+     */
     public function actionRenameCategory($moduleId, $oldCategory, $newCategory)
     {
         RenameTranslationCategory::rename($moduleId, $oldCategory, $newCategory);